requirements.txt
numpy
faiss-cpu
requests
sentence-transformers
python-dotenv
